Quick benchmarks (first line is the original code, second is the patched one):

```shell
$ python -m timeit -n 1000000 -s "from hexbytes import HexBytes" "HexBytes(0)"
1000000 loops, best of 5: 1.260 usec per loop
1000000 loops, best of 5: 0.875 nsec per loop (-30.5 %)

$ python -m timeit -n 1000000 -s "from hexbytes import HexBytes" "HexBytes('0')"
1000000 loops, best of 5: 883 nsec per loop
1000000 loops, best of 5: 760 nsec per loop (-13.9 %)

$ python -m timeit -n 1000000 -s "from hexbytes import HexBytes" "HexBytes(False)"
1000000 loops, best of 5: 540 nsec per loop
1000000 loops, best of 5: 447 nsec per loop (-17.2 %)
```

## What was wrong?

Nothing.

## How was it fixed?

Gut feeling.

I had to bypass mypy errors as they are not revelant.

### To-Do

[//]: # (Stay ahead of things, add list items here!)
- [x] Clean up commit history

[//]: # (For important changes that should go into the release notes please add a newsfragment file as explained here: https://github.com/ethereum/hexbytes/blob/master/newsfragments/README.md)

[//]: # (See: https://hexbytes.readthedocs.io/en/latest/contributing.html#pull-requests)
- [ ] Add entry to the [release notes](https://github.com/ethereum/hexbytes/blob/master/newsfragments/README.md)

#### Cute Animal Picture

![put a cute animal picture link inside the parentheses](https://4.bp.blogspot.com/-8lDc8tk0ip8/URNZFUbT7iI/AAAAAAAAAes/oul3yNGRkq8/s1600/Cute-LIttle-Monkey-HD-Wallpaper.jpg)